# lib/api-zod/src/generated/types/dnsRecord.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DnsRecord {
  /**
     * Resend's record grouping, e.g. SPF or DKIM
     * @nullable
     */
  record?: string | null;
  type: string;
  name: string;
  value: string;
  /** @nullable */
  ttl?: string | null;
  /** @nullable */
  priority?: number | null;
  /** @nullable */
  status?: string | null;
}