# lib/api-zod/src/generated/types/issue.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Issue {
  id: number;
  subject: string;
  sentAt: Date;
  recipientCount: number;
  failedCount: number;
}