# lib/api-zod/src/generated/types/issueSendInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface IssueSendInput {
  /**
     * @minLength 1
     * @maxLength 500
     */
  subject: string;
  /** @minLength 1 */
  body: string;
  confirmDuplicate?: boolean;
}