# lib/api-zod/src/generated/types/sendResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SendError } from './sendError';

export interface SendResult {
  sent: number;
  failed: number;
  errors: SendError[];
}