# lib/api-zod/src/generated/types/subscriberInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SubscriberInput {
  /** @maxLength 320 */
  email: string;
  /** @maxLength 200 */
  firstName?: string;
}